+++ Example/KinescopeExample/Flows/VideoViewController.swift
@@ -2,6 +2,12 @@ import UIKit
 import KinescopeSDK
 
 final class VideoViewController: UIViewController {
+    
+    // MARK: - Nested Types
+
+    private enum CustomPlayerOption: String {
+        case share
+    }
 
     // MARK: - IBOutlets
 
@@ -40,6 +46,15 @@ final class VideoViewController: UIViewController {
         PipManager.shared.closePipIfNeeded(with: videoId)
 
         player = KinescopeVideoPlayer(config: .init(videoId: videoId))
+
+        if #available(iOS 13.0, *) {
+            if let shareIcon = UIImage(systemName: "square.and.arrow.up")?.withRenderingMode(.alwaysTemplate) {
+                player?.addCustomPlayerOption(with: CustomPlayerOption.share, and: shareIcon)
+            }
+        }
+        player?.disableOptions([.airPlay])
+        
+        player?.setDelegate(delegate: self)
         player?.attach(view: playerView)
         player?.play()
         player?.pipDelegate = PipManager.shared
@@ -52,4 +67,25 @@ extension VideoViewController: UINavigationControllerDelegate {
     func navigationControllerSupportedInterfaceOrientations(_ navigationController: UINavigationController) -> UIInterfaceOrientationMask {
         return self.supportedInterfaceOrientations
     }
+}
+
+extension VideoViewController: KinescopeVideoPlayerDelegate {
+
+    func player(didSelectCustomOptionWith optionId: AnyHashable, anchoredAt view: UIView) {
+        guard let option = optionId as? CustomPlayerOption else {
+            return
+        }
+
+        switch option {
+        case .share:
+            let items = [player?.config.shareLink]
+            let activityViewController = UIActivityViewController(activityItems: items as [Any], 
+                                                                  applicationActivities: nil)
+            let presentationController = activityViewController.presentationController as? UIPopoverPresentationController
+            presentationController?.sourceView = view
+            presentationController?.permittedArrowDirections = .down
+            present(activityViewController, animated: true)
+        }
+    }
+
 }