+++ Sources/KinescopeSDK/Player/KinescopeVideoPlayerDelegate.swift
@@ -38,6 +38,8 @@ public protocol KinescopeVideoPlayerDelegate: AnyObject {
     func player(didFastBackwardTo time: TimeInterval)
     /// Triggered on quality change
     func player(changedQualityTo quality: String)
+    /// Triggered on custom option button selected in options menu
+    func player(didSelectCustomOptionWith optionId: AnyHashable, anchoredAt view: UIView)
 }
 
 public extension KinescopeVideoPlayerDelegate {