+++ Sources/KinescopeSDK/View/Player/KinescopePlayerView.swift
@@ -496,6 +496,8 @@ extension KinescopePlayerView: PlayerControlOutput {
         case .pip:
             let isPipActive = pipController?.isPictureInPictureActive ?? false
             isPipActive ? pipController?.stopPictureInPicture() : pipController?.startPictureInPicture()
+        case .custom(let id, _):
+            delegate?.didSelect(option: id, optionView: controlPanel?.getCustomOptionView(by: id) ?? self)
         default:
             break
         }