+++ Sources/KinescopeSDK/View/Player/KinescopePlayerViewDelegate.swift
@@ -12,4 +12,5 @@ protocol KinescopePlayerViewDelegate: AnyObject {
     func didSelectAttachment(with id: String)
     func didSelectDownloadAll(for title: String)
     func didSelect(subtitles: String)
+    func didSelect(option: AnyHashable, optionView: UIView)
 }