+++ CHANGELOG.md
@@ -16,4 +16,4 @@
 - live stream announce view
 - collecting of analytic events with playback data for Kinescope dashboard
 - textField in example project to allow user to input videoId
-
+- ability to manage options menu on player view